// src/app/core/services/models/mission-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MissionRequest {
  description: string;
  endDate: string;
  requiredSkill: 'COMMUNICATION' | 'ORGANIZATION' | 'IT' | 'TEACHING' | 'FIRST_AID' | 'LOGISTICS' | 'PSYCHOLOGICAL_SUPPORT' | 'LEADERSHIP';
  startDate: string;
  title: string;
}