// src/app/core/services/models/mission-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MissionResponse {
  creationDate?: string;
  description?: string;
  endDate?: string;
  id?: number;
  isAsseigned?: boolean;
  lastModifiedDate?: string;
  requiredSkill?: 'COMMUNICATION' | 'ORGANIZATION' | 'IT' | 'TEACHING' | 'FIRST_AID' | 'LOGISTICS' | 'PSYCHOLOGICAL_SUPPORT' | 'LEADERSHIP';
  startDate?: string;
  status?: 'PENDING' | 'IN_PROGRESS' | 'COMPLETED' | 'CANCELLATION';
  title?: string;
}